Release beta (#2139)
* Refactor pod (#2134)

* refactor pod

* T29113 - Fix Push Notification (#2133)

* add deeplink to checkout

* go to payment

* fix observable + error

* fix error message

* add activity indicator

* change error from pop to no result

* fix error view

* fix crash on purchase

* clean code

* fix code review

* fix code review 2.0

* fix force unwrap

* add dispatch to trackAppLocation

* update pod

* Air panas (#2135)

* restructure folder, move third party library, move constant string

* Move Component & Helper to new folder

* remove unnecessary file

* move some code from cocoa folder

* refactor some helper

* change component folder into common

* compile all extension in one file & add some useful extension

* Combine some swift extension into one file

* Xcode Build Optimization

* remove veritrans & add some tweak

* update JLPermission pod

* Complete podcast version for view component

* give current version in existing podcast

* update DKImagePickerController

* update Flex Version

* update Eureka version

* update DownPicker

* after merge

* Tydiing up podfile

* object refactor part 1

* Object refactor part 2

* refactor Activation folder

* refactor Address folder

* refactor Alert folder

* refactor camera folder

* refactor Catalog folder

* refactor Category folder

* refactor Controller folder

* refactor CustomViews folder

* refactor Deposit folder

* refactor Forgot Password folder & some xib

* refactor GeneralCell folder

* refactor GeneralViewController folder

* refactor Home folder

* refactor Hotlist folder

* refactor Inbox Message folder

* refactor InboxResolutionCenter folder

* refactor InboxTalk folder

* refactor InboxTicket folder

* refactor More folder

* refactor MyShopAddress folder

* refactor MyShopEtalase folder

* refactor MyShopNote folder

* refactor MyShopPayment folder

* refactor MyShopShipment folder

* refactor navigator folder

* refactor NotificationManager folder

* refactor onboarding folder

* refactor PriceAlert folder

* refactor product folder

* refactor promo folder

* refactor ResolutionCenter folder

* refactor ReusableView folder

* refactor Sales folder

* refactor Search folder

* refactor Settings folder

* refactor Shop folder

* refactor ShopPageView folder

* refactor SortFilterShare folder

* refactor Talk folder

* refactor TokopediaNetworkManager folder

* refactor TrackingOrder folder

* refactor Transaction folder

* refactor TXOrder folder

* refactor UserPageView folder

* refactor Sales folder

* refactor remaining warning

* refactor masonry warning

* get temporary restructure

* restore temporary file

* fix merge conflict

* refactor podfile

* T29113 - Fix Push Notification (#2133)

* add deeplink to checkout

* go to payment

* fix observable + error

* fix error message

* add activity indicator

* change error from pop to no result

* fix error view

* fix crash on purchase

* clean code

* fix code review

* fix code review 2.0

* fix force unwrap

* add dispatch to trackAppLocation

* restore pod

* updating some code review feedback

+++ ios/Tokopedia/Classses/UIRefactor/CustomViews/ScrollViewSimultaneousGesture.m
@@ -23,7 +23,7 @@
 - (BOOL)gestureRecognizerShouldBegin:(UIGestureRecognizer *)gestureRecognizer {
     if ([gestureRecognizer isKindOfClass:[UIPanGestureRecognizer class]]) {
         CGPoint velocity = [(UIPanGestureRecognizer *)gestureRecognizer velocityInView:self];
-        if (fabsf(velocity.y) * 2 < fabsf(velocity.x)) {
+        if (fabs(velocity.y) * 2 < fabs(velocity.x)) {
             return YES;
         }
     }
